requirements.txt
bidict==0.22.0
certifi==2022.6.15
charset-normalizer==2.1.1
eventlet>=0.33.1
Flask>=2.2.0
Flask-SocketIO>=5.2.0
idna==3.3
python-engineio==4.3.4
python-socketio==5.7.1
requests==2.28.1
sortedcontainers>=2.4.0
urllib3==1.26.12
websocket-client==1.4.0
